Cannot reproduce infinite loop bug on this commit.
I believe the infinite loop was caused from drowning the webview in "Task.Run" invocations.

So for example, holding shift so one can scroll horizontally with the mouse wheel was the most common case for what seemed to cause freezing in the app.

When holding shift I saw that I was firing off a ton of keyboard events and going through much code all for nothing.

So for shift keys being held down they are ignored, and other changes as well.

I have way too much throttling logic going on in this commit, and foreach throttle its TimeSpan is way to long. But I was doing so to hunt down this infinite loop bug.

I'm going to mess about further after this gets committed. Because right now the throttling is definitely overkill.